[RUI-251]: Move placement of types
