+++ src/components/charts/scatter/BubbleChartPro/BubbleChartPro.types.ts
@@ -1,0 +1,5 @@
+import type { ScatterChartProOptionsClickArg } from '../ScatterChartPro/ScatterChartPro.types';
+
+export type BubbleChartProOptionsClickArg = ScatterChartProOptionsClickArg & {
+  sizeMeasureValue: string;
+};
